+++ src/Element/LoqatePcaAddress.php
@@ -2,7 +2,11 @@
 
 namespace Drupal\loqate\Element;
 
+use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Render\Element\FormElement;
+use Drupal\Core\StringTranslation\TranslatableMarkup;
+use Drupal\loqate\PcaAddressFieldMapping\PcaAddressElement;
+use Drupal\loqate\PcaAddressTrait;
 
 /**
  * Provides a standalone Loqate PCA address form element.
@@ -38,14 +42,65 @@ use Drupal\Core\Render\Element\FormElement;
  */
 class LoqatePcaAddress extends FormElement {
 
+  use PcaAddressTrait;
+
   /**
    * {@inheritdoc}
    */
   public function getInfo() {
     $class = get_class($this);
     return [
-
+      '#process' => [
+        [$class, 'processAddress'],
+        [$class, 'processPcaAddress'],
+      ],
+      '#pca_fields' => [],
+      '#pca_options' => [],
+      '#show_address_fields' => FALSE,
+      '#allow_manual_input' => TRUE,
+      '#attached' => [
+        // @TODO: Move this to base module.
+        'library' => ['pca_address/element.pca_address.address.js'],
+      ],
     ];
+  }
+
+  /**
+   * Process the address fields.
+   */
+  public static function processAddress(array &$element, FormStateInterface $form_state, array &$complete_form) {
+
+    $element[PcaAddressElement::LINE1] = [
+      '#type' => 'textfield',
+      '#title' => new TranslatableMarkup('Address Line 1'),
+    ];
+
+    $element[PcaAddressElement::LINE2] = [
+      '#type' => 'textfield',
+      '#title' => new TranslatableMarkup('Address Line 2'),
+    ];
+
+    $element[PcaAddressElement::LOCALITY] = [
+      '#type' => 'textfield',
+      '#title' => new TranslatableMarkup('City/Town'),
+    ];
+
+    $element[PcaAddressElement::ADMINISTRATIVE_AREA] = [
+      '#type' => 'textfield',
+      '#title' => new TranslatableMarkup('State/Province'),
+    ];
+
+    $element[PcaAddressElement::POSTAL_CODE] = [
+      '#type' => 'textfield',
+      '#title' => new TranslatableMarkup('ZIP/Postal Code'),
+    ];
+
+    $element[PcaAddressElement::COUNTRY_CODE] = [
+      '#type' => 'textfield',
+      '#title' => new TranslatableMarkup('Country'),
+    ];
+
+    return $element;
   }
 
 }
